Refactor Task class into subclasses and add method to show their format in data file
