+++ src/main/java/TodoTask.java
@@ -1,0 +1,7 @@
+import java.util.Map;
+
+public class TodoTask extends Task {
+    public TodoTask(Map<String, String> args, boolean isDone) {
+        super('T', args.get("description"), isDone);
+    }
+}
